@@ -1,6 +1,7 @@
 """Tests for MoogoClient schedule operations."""
 
-from typing import Any, Callable
+from collections.abc import Callable
+from typing import Any
 from unittest.mock import AsyncMock, MagicMock
 
 import pytest
